chore: move env related code
